requirements.txt
transformers
openai
torch
Pillow
python-telegram-bot==20.3
